pyproject.toml
[project]
name = "scholar-rss"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "openai>=1.0.0",
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
    "ipdb>=0.13.13",
]
